+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to `omines\antispam-bundle` will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+ - `FORM_PROCESSED` event allows processing ham as well as spam
+ - `quarantine.only_spam` configuration enables storing only spam or both ham and spam in quarantine
+
+### Breaking
+ - `FormViolationEvent` renamed to `FormResultEvent`
 
 ## [0.1.4] - 2023-11-17
 ### Added